pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "sharkpoint"
version = "1.0.0"
description = "A small Pythonic library for interacting with SharePoint document libraries"
authors = [
  { name = "TheOtherOne" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.7"
dependencies = [
  "requests >=2.6",
  "azure-identity",
]

[project.urls]
"Homepage" = "https://github.com/ThatsItForTheOtherOne/sharkpoint"
"Bug Tracker" = "https://github.com/ThatsItForTheOtherOne/sharkpoint/issues"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
  "src"
]
env_files = ".env_tests"